+++ go-controller/pkg/node/healthcheck_service.go
@@ -1,0 +1,223 @@
+package node
+
+import (
+	"fmt"
+	"net"
+	"sync"
+
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/config"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/factory"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/kube/healthcheck"
+
+	kapi "k8s.io/api/core/v1"
+	discovery "k8s.io/api/discovery/v1"
+	ktypes "k8s.io/apimachinery/pkg/types"
+	apierrors "k8s.io/apimachinery/pkg/util/errors"
+	"k8s.io/apimachinery/pkg/util/sets"
+	utilnet "k8s.io/utils/net"
+)
+
+// initLoadBalancerHealthChecker initializes the health check server for
+// ServiceTypeLoadBalancer services
+
+type loadBalancerHealthChecker struct {
+	sync.Mutex
+	nodeName     string
+	server       healthcheck.Server
+	services     map[ktypes.NamespacedName]uint16
+	endpoints    map[ktypes.NamespacedName]int
+	watchFactory factory.NodeWatchFactory
+}
+
+func newLoadBalancerHealthChecker(nodeName string, watchFactory factory.NodeWatchFactory) *loadBalancerHealthChecker {
+	return &loadBalancerHealthChecker{
+		nodeName:     nodeName,
+		server:       healthcheck.NewServer(nodeName, nil, nil, nil),
+		services:     make(map[ktypes.NamespacedName]uint16),
+		endpoints:    make(map[ktypes.NamespacedName]int),
+		watchFactory: watchFactory,
+	}
+}
+
+func (l *loadBalancerHealthChecker) AddService(svc *kapi.Service) error {
+	if svc.Spec.HealthCheckNodePort != 0 {
+		l.Lock()
+		defer l.Unlock()
+		name := ktypes.NamespacedName{Namespace: svc.Namespace, Name: svc.Name}
+		l.services[name] = uint16(svc.Spec.HealthCheckNodePort)
+		return l.server.SyncServices(l.services)
+	}
+	return nil
+}
+
+func (l *loadBalancerHealthChecker) UpdateService(old, new *kapi.Service) error {
+	// if the ETP values have changed between local and cluster,
+	// we need to update health checks accordingly
+	// HealthCheckNodePort is used only in ETP=local mode
+	var err error
+	var errors []error
+	if old.Spec.ExternalTrafficPolicy == kapi.ServiceExternalTrafficPolicyTypeCluster &&
+		new.Spec.ExternalTrafficPolicy == kapi.ServiceExternalTrafficPolicyTypeLocal {
+		if err = l.AddService(new); err != nil {
+			errors = append(errors, err)
+		}
+		epSlices, err := l.watchFactory.GetEndpointSlices(new.Namespace, new.Name)
+		if err != nil {
+			errors = append(errors, fmt.Errorf("could not fetch endpointslices "+
+				"for service %s/%s during health check update service: %v",
+				new.Namespace, new.Name, err))
+			return apierrors.NewAggregate(errors)
+		}
+		namespacedName := ktypes.NamespacedName{Namespace: new.Namespace, Name: new.Name}
+		l.Lock()
+		l.endpoints[namespacedName] = l.CountLocalEndpointAddresses(epSlices)
+		if err = l.server.SyncEndpoints(l.endpoints); err != nil {
+			errors = append(errors, err)
+		}
+		l.Unlock()
+	}
+	if old.Spec.ExternalTrafficPolicy == kapi.ServiceExternalTrafficPolicyTypeLocal &&
+		new.Spec.ExternalTrafficPolicy == kapi.ServiceExternalTrafficPolicyTypeCluster {
+		if err = l.DeleteService(old); err != nil {
+			errors = append(errors, err)
+		}
+	}
+	return apierrors.NewAggregate(errors)
+}
+
+func (l *loadBalancerHealthChecker) DeleteService(svc *kapi.Service) error {
+	if svc.Spec.HealthCheckNodePort != 0 {
+		l.Lock()
+		defer l.Unlock()
+		name := ktypes.NamespacedName{Namespace: svc.Namespace, Name: svc.Name}
+		delete(l.services, name)
+		delete(l.endpoints, name)
+		return l.server.SyncServices(l.services)
+	}
+	return nil
+}
+
+func (l *loadBalancerHealthChecker) SyncServices(svcs []interface{}) error {
+	return nil
+}
+
+func (l *loadBalancerHealthChecker) SyncEndPointSlices(epSlice *discovery.EndpointSlice) error {
+	namespacedName, err := namespacedNameFromEPSlice(epSlice)
+	if err != nil {
+		return fmt.Errorf("skipping %s/%s: %v", epSlice.Namespace, epSlice.Name, err)
+	}
+	epSlices, err := l.watchFactory.GetEndpointSlices(epSlice.Namespace, epSlice.Labels[discovery.LabelServiceName])
+	if err != nil {
+		// should be a rare occurence
+		return fmt.Errorf("could not fetch all endpointslices for service %s during health check", namespacedName.String())
+	}
+
+	localEndpointAddressCount := l.CountLocalEndpointAddresses(epSlices)
+	if len(epSlices) == 0 {
+		// let's delete it from cache and wait for the next update;
+		// this will show as 0 endpoints for health checks
+		delete(l.endpoints, namespacedName)
+	} else {
+		l.endpoints[namespacedName] = localEndpointAddressCount
+	}
+	return l.server.SyncEndpoints(l.endpoints)
+}
+
+func (l *loadBalancerHealthChecker) AddEndpointSlice(epSlice *discovery.EndpointSlice) error {
+	namespacedName, err := namespacedNameFromEPSlice(epSlice)
+	if err != nil {
+		return fmt.Errorf("cannot add %s/%s to loadBalancerHealthChecker: %v", epSlice.Namespace, epSlice.Name, err)
+	}
+	l.Lock()
+	defer l.Unlock()
+	if _, exists := l.services[namespacedName]; exists {
+		return l.SyncEndPointSlices(epSlice)
+	}
+	return nil
+}
+
+func (l *loadBalancerHealthChecker) UpdateEndpointSlice(oldEpSlice, newEpSlice *discovery.EndpointSlice) error {
+	namespacedName, err := namespacedNameFromEPSlice(newEpSlice)
+	if err != nil {
+		return fmt.Errorf("cannot update %s/%s in loadBalancerHealthChecker: %v",
+			newEpSlice.Namespace, newEpSlice.Name, err)
+	}
+	l.Lock()
+	defer l.Unlock()
+	if _, exists := l.services[namespacedName]; exists {
+		return l.SyncEndPointSlices(newEpSlice)
+	}
+	return nil
+}
+
+func (l *loadBalancerHealthChecker) DeleteEndpointSlice(epSlice *discovery.EndpointSlice) error {
+	l.Lock()
+	defer l.Unlock()
+	return l.SyncEndPointSlices(epSlice)
+}
+
+// CountLocalEndpointAddresses returns the number of IP addresses from ready endpoints that are local
+// to the node for a service
+func (l *loadBalancerHealthChecker) CountLocalEndpointAddresses(endpointSlices []*discovery.EndpointSlice) int {
+	localEndpointAddresses := sets.NewString()
+	for _, endpointSlice := range endpointSlices {
+		for _, endpoint := range endpointSlice.Endpoints {
+			isLocal := endpoint.NodeName != nil && *endpoint.NodeName == l.nodeName
+			if isEndpointReady(endpoint) && isLocal {
+				localEndpointAddresses.Insert(endpoint.Addresses...)
+			}
+		}
+	}
+	return len(localEndpointAddresses)
+}
+
+// hasLocalHostNetworkEndpoints returns true if there is at least one host-networked endpoint
+// in the provided list that is local to this node.
+// It returns false if none of the endpoints are local host-networked endpoints or if ep.Subsets is nil.
+func hasLocalHostNetworkEndpoints(epSlices []*discovery.EndpointSlice, nodeAddresses []net.IP) bool {
+	for _, epSlice := range epSlices {
+		for _, endpoint := range epSlice.Endpoints {
+			if !isEndpointReady(endpoint) {
+				continue
+			}
+			for _, ip := range endpoint.Addresses {
+				for _, nodeIP := range nodeAddresses {
+					if nodeIP.String() == utilnet.ParseIPSloppy(ip).String() {
+						return true
+					}
+				}
+			}
+		}
+	}
+	return false
+}
+
+// isHostEndpoint determines if the given endpoint ip belongs to a host networked pod
+func isHostEndpoint(endpointIP string) bool {
+	for _, clusterNet := range config.Default.ClusterSubnets {
+		if clusterNet.CIDR.Contains(net.ParseIP(endpointIP)) {
+			return false
+		}
+	}
+	return true
+}
+
+// discovery.EndpointSlice reports in the same slice all endpoints along with their status.
+// isEndpointReady takes an endpoint from an endpoint slice and returns true if the endpoint is
+// to be considered ready.
+func isEndpointReady(endpoint discovery.Endpoint) bool {
+	return endpoint.Conditions.Ready == nil || *endpoint.Conditions.Ready
+}
+
+func namespacedNameFromEPSlice(epSlice *discovery.EndpointSlice) (ktypes.NamespacedName, error) {
+	// Return the namespaced name of the corresponding service
+	var serviceNamespacedName ktypes.NamespacedName
+	svcName := epSlice.Labels[discovery.LabelServiceName]
+	if svcName == "" {
+		// should not happen, since the informer already filters out endpoint slices with an empty service label
+		return serviceNamespacedName,
+			fmt.Errorf("endpointslice %s/%s: empty value for label %s",
+				epSlice.Namespace, epSlice.Name, discovery.LabelServiceName)
+	}
+	return ktypes.NamespacedName{Namespace: epSlice.Namespace, Name: svcName}, nil
+}
